\name{abcs-package}
\alias{abcs-package}
\alias{abcs}
\docType{package}
\title{
\packageTitle{abcs}
}
\description{
\packageDescription{abcs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{abcs}
\packageIndices{abcs}

}
\author{
\packageAuthor{abcs}

Maintainer: \packageMaintainer{abcs}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
